requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.7.1
Django==5.1.1
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
elementpath==4.5.0
idna==3.10
isodate==0.6.1
joblib==1.4.2
lxml==5.3.0
numpy==2.1.1
pandas==2.2.3
psycopg2==2.9.9
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.2.1
pysaml2==7.5.0
python-dateutil==2.9.0.post0
python3-saml==1.16.0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xmlschema==2.5.1
xmlsec==1.3.14
